refactor(#741): refactor grottoController
